// weeb-dashboard/lib/plugins/myanimelist/types.d.ts
export interface MyAnimeListConfig {
    enabled: boolean;
    sections: string[];
    username: string;
    favorites_max?: number;
}
export interface AnimeStatistics {
    days_watched: number;
    mean_score: number;
    watching: number;
    completed: number;
    on_hold: number;
    dropped: number;
    plan_to_watch: number;
    total_entries: number;
    rewatched: number;
    episodes_watched: number;
}
export interface MangaStatistics {
    days_read: number;
    mean_score: number;
    reading: number;
    completed: number;
    on_hold: number;
    dropped: number;
    plan_to_read: number;
    total_entries: number;
    reread: number;
    chapters_read: number;
    volumes_read: number;
}
export interface LastUpdatesAnime {
    title: string;
    image: string;
    score: number;
    status: string;
    episodes_seen: number | null;
    episodes_total: number | null;
    date: string;
}
export interface LastUpdatesManga {
    title: string;
    image: string;
    score: number;
    status: string;
    chapters_read: number | null;
    chapters_total: number | null;
    date: string;
}
export interface MalLastUpdates {
    anime: LastUpdatesAnime[];
    manga: LastUpdatesManga[];
}
export interface MyAnimeListData {
    favorites: {
        anime: BasicAnimeFavorite[];
        manga: BasicMangaFavorite[];
        characters: BasicCharacterFavorite[];
        people: BasicPeopleFavorite[];
    };
    favorites_full: {
        anime: FullAnimeFavorite[];
        manga: FullMangaFavorite[];
        characters: BasicCharacterFavorite[];
        people: BasicPeopleFavorite[];
    };
    last_updated: MalLastUpdates;
    statistics: {
        anime: AnimeStatistics;
        manga: MangaStatistics;
    };
}
export interface BasicAnimeFavorite {
    mal_id: number;
    title: string;
    image: string;
    start_year: number;
    type: string;
}
export interface FullAnimeFavorite extends BasicAnimeFavorite {
    synopsis: string;
    score: number;
    popularity: number;
    episodes: number | null;
    status: string;
    rank: number;
    year: number | null;
    genres?: {
        name: string;
    }[];
}
export interface BasicMangaFavorite {
    mal_id: number;
    title: string;
    image: string;
    start_year: number;
    type: string;
}
export interface FullMangaFavorite extends BasicMangaFavorite {
    synopsis: string;
    score: number;
    popularity: number;
    chapters: number | null;
    rank: number;
    volumes: number | null;
    status: string;
    year: number | null;
    genres?: {
        name: string;
    }[];
}
export interface BasicCharacterFavorite {
    mal_id: number;
    name: string;
    image: string;
}
export interface BasicPeopleFavorite {
    mal_id: number;
    name: string;
    image: string;
}